Further simplify ReplayCaptureController
