+++ platform/jvm/gradle-test-app/src/androidTest/java/io/bitdrift/gradletestapp/TestUtils.kt
@@ -67,7 +67,6 @@ object TestUtils {
                     Log.e("Replay Tests", message, e)
                 }
             },
-            SessionReplayConfiguration(),
             context,
         )
     }
